requirements.txt
numpy
sklearn
tensorflow-gpu==1.15.0
keras==2.1.2
